# Complete API Performance Optimization: Production-Ready with Cache Invalidation