--- a/includes/class-hrw-json-timing.php
+++ b/includes/class-hrw-json-timing.php
@@ -1,10 +1,10 @@
 <?php
 
 /**
- * HRW JSON Timing Class
+ * HRW JSON Timing Measurement
  * 
- * Intercepts WordPress JSON encoding to measure serialization time
- * This helps identify if JSON encoding is the 5-second bottleneck
+ * Measures the time taken for JSON encoding of API responses
+ * Helps identify if JSON serialization is a performance bottleneck
  * 
  * @package HRW_Plugin
  * @since 1.2.0
@@ -18,12 +18,12 @@ if (!defined('ABSPATH')) {
 class HRW_JSON_Timing
 {
 	/**
-	 * Track if we've already hooked into json encoding
+	 * Track if hooks are already initialized to prevent duplicate hooks
 	 */
 	private static $hooked = false;
 
 	/**
-	 * Store timing data
+	 * Store timing data for analysis
 	 */
 	private static $timing_data = [];
 
@@ -32,73 +32,75 @@ class HRW_JSON_Timing
 	 */
 	public static function init()
 	{
+		// Prevent duplicate initialization
 		if (self::$hooked) {
 			return;
 		}
 
-		// Hook into REST API response before it gets JSON encoded
+		// Hook into REST response to measure JSON encoding
 		add_filter('rest_post_dispatch', [__CLASS__, 'time_json_encoding'], 10, 3);
 
 		self::$hooked = true;
-		error_log('HRW JSON Timing: Initialized JSON timing hooks');
+
+		// Only log initialization in debug mode
+		if (defined('WP_DEBUG') && WP_DEBUG) {
+			error_log('HRW JSON Timing: Initialized JSON timing hooks');
+		}
 	}
 
 	/**
-	 * Time the JSON encoding process
+	 * Measure JSON encoding time for VibeMap API responses
 	 * 
-	 * @param WP_HTTP_Response $result  Result to send to the client.
-	 * @param WP_REST_Server   $server  Server instance.
-	 * @param WP_REST_Request  $request Request used to generate the response.
-	 * @return WP_HTTP_Response
+	 * @param WP_REST_Response $result REST response object
+	 * @param WP_REST_Server $server REST server instance
+	 * @param WP_REST_Request $request REST request object
+	 * @return WP_REST_Response Modified response with timing data
 	 */
 	public static function time_json_encoding($result, $server, $request)
 	{
-		// Only time our specific endpoint
+		// Only measure our specific endpoint
 		if ($request->get_route() !== '/vibemap/v1/places-data') {
 			return $result;
 		}
 
-		error_log('HRW JSON Timing: [TIMING] Starting JSON encoding measurement at ' . date('H:i:s.') . substr(microtime(), 2, 3));
-
-		// Get the data that will be JSON encoded
-		$response_data = $result->get_data();
-
-		// Time JSON encoding
+		// Start timing measurement
 		$json_start = microtime(true);
+		$response_data = $result->get_data();
 		$json_string = json_encode($response_data);
 		$json_time = round((microtime(true) - $json_start) * 1000, 2);
 
-		// Calculate payload size
+		// Calculate payload information
 		$payload_size = strlen($json_string);
 		$payload_mb = round($payload_size / 1024 / 1024, 2);
+		$restaurant_count = count($response_data['places'] ?? []);
 
-		// Log detailed timing information
-		error_log('HRW JSON Timing: [TIMING] JSON encoding took ' . $json_time . 'ms');
-		error_log('HRW JSON Timing: [TIMING] Payload size: ' . $payload_mb . 'MB (' . number_format($payload_size) . ' bytes)');
-		error_log('HRW JSON Timing: [TIMING] Average per restaurant: ' . round($json_time / 252, 2) . 'ms');
-
-		// Store timing data for potential optimization
+		// Store timing data
 		self::$timing_data[] = [
 			'timestamp' => time(),
 			'json_time_ms' => $json_time,
 			'payload_size_bytes' => $payload_size,
 			'payload_size_mb' => $payload_mb,
-			'restaurant_count' => count($response_data['places'] ?? []),
+			'restaurant_count' => $restaurant_count,
 		];
 
-		// Add timing info to response data for debugging
+		// Add timing info to debug data if available
 		if (isset($response_data['debug_info'])) {
 			$response_data['debug_info']['json_encoding_ms'] = $json_time;
 			$response_data['debug_info']['payload_size_mb'] = $payload_mb;
 			$response_data['debug_info']['total_backend_time_ms'] =
 				($response_data['debug_info']['timing']['data_merger_ms'] ?? 0) + $json_time;
-
-			// Update the result with timing info
 			$result->set_data($response_data);
 		}
 
-		error_log('HRW JSON Timing: [TIMING] JSON encoding measurement complete at ' . date('H:i:s.') . substr(microtime(), 2, 3));
-		error_log('HRW JSON Timing: [TIMING] ===== JSON TIMING COMPLETE =====');
+		// Only log detailed timing in debug mode
+		if (defined('WP_DEBUG') && WP_DEBUG) {
+			error_log('HRW JSON Timing: JSON encoding took ' . $json_time . 'ms for ' . $payload_mb . 'MB payload');
+		}
+
+		// Log performance issues even in production
+		if ($json_time > 100) { // More than 100ms is concerning
+			error_log('HRW JSON Timing: SLOW JSON encoding detected: ' . $json_time . 'ms for ' . $payload_mb . 'MB');
+		}
 
 		return $result;
 	}
@@ -111,21 +113,18 @@ class HRW_JSON_Timing
 	public static function get_timing_stats()
 	{
 		if (empty(self::$timing_data)) {
-			return [
-				'message' => 'No timing data collected yet',
-				'suggestion' => 'Make an API call to /vibemap/v1/places-data to collect timing data'
-			];
+			return [];
 		}
 
-		$latest = end(self::$timing_data);
-		$all_times = array_column(self::$timing_data, 'json_time_ms');
+		$recent_data = array_slice(self::$timing_data, -10); // Last 10 measurements
+		$total_time = array_sum(array_column($recent_data, 'json_time_ms'));
+		$avg_time = round($total_time / count($recent_data), 2);
 
 		return [
-			'latest_call' => $latest,
-			'average_json_time_ms' => round(array_sum($all_times) / count($all_times), 2),
-			'min_json_time_ms' => min($all_times),
-			'max_json_time_ms' => max($all_times),
-			'total_calls_measured' => count(self::$timing_data),
+			'recent_measurements' => count($recent_data),
+			'average_json_time_ms' => $avg_time,
+			'total_measurements' => count(self::$timing_data),
+			'last_measurement' => end($recent_data)
 		];
 	}
 
@@ -135,6 +134,39 @@ class HRW_JSON_Timing
 	public static function clear_timing_data()
 	{
 		self::$timing_data = [];
-		error_log('HRW JSON Timing: Cleared all timing data');
+
+		if (defined('WP_DEBUG') && WP_DEBUG) {
+			error_log('HRW JSON Timing: Cleared all timing data');
+		}
+	}
+
+	/**
+	 * Get performance report for admin
+	 * 
+	 * @return array Performance report
+	 */
+	public static function get_performance_report()
+	{
+		$stats = self::get_timing_stats();
+
+		if (empty($stats)) {
+			return ['status' => 'No timing data available'];
+		}
+
+		$status = 'Good';
+		if ($stats['average_json_time_ms'] > 50) {
+			$status = 'Needs optimization';
+		} elseif ($stats['average_json_time_ms'] > 20) {
+			$status = 'Acceptable';
+		}
+
+		return [
+			'status' => $status,
+			'average_time' => $stats['average_json_time_ms'] . 'ms',
+			'measurements' => $stats['recent_measurements'],
+			'recommendation' => $stats['average_json_time_ms'] > 50
+				? 'Consider payload size reduction'
+				: 'JSON performance is optimal'
+		];
 	}
 }