--- a/includes/class-hrw-wordpress-timing.php
+++ b/includes/class-hrw-wordpress-timing.php
@@ -1,10 +1,10 @@
 <?php
 
 /**
- * HRW WordPress Timing Class
+ * HRW WordPress Lifecycle Timing
  * 
- * Measures WordPress-level timing to identify overhead after our backend completes
- * This helps find the gap between our 13ms backend and 5.57s browser timing
+ * Measures WordPress request lifecycle to identify hosting infrastructure delays
+ * Tracks init, wp_loaded, rest_serve, and shutdown phases
  * 
  * @package HRW_Plugin
  * @since 1.2.0
@@ -18,7 +18,7 @@ if (!defined('ABSPATH')) {
 class HRW_WordPress_Timing
 {
 	/**
-	 * Track if we've already hooked
+	 * Track if hooks are already initialized
 	 */
 	private static $hooked = false;
 
@@ -32,49 +32,64 @@ class HRW_WordPress_Timing
 	 */
 	public static function init()
 	{
+		// Prevent duplicate initialization
 		if (self::$hooked) {
 			return;
 		}
 
-		// Hook early in WordPress request lifecycle
+		// Track WordPress lifecycle phases
 		add_action('init', [__CLASS__, 'track_request_start'], 1);
-
-		// Hook late in WordPress response
 		add_action('wp_loaded', [__CLASS__, 'track_wp_loaded'], 999);
 		add_action('shutdown', [__CLASS__, 'track_request_end'], 999);
-
-		// Hook into REST API specifically
 		add_filter('rest_pre_serve_request', [__CLASS__, 'track_rest_serve'], 10, 4);
 
 		self::$hooked = true;
-		error_log('HRW WordPress Timing: Initialized WordPress timing hooks');
+
+		// Only log initialization in debug mode
+		if (defined('WP_DEBUG') && WP_DEBUG) {
+			error_log('HRW WordPress Timing: Initialized WordPress timing hooks');
+		}
 	}
 
 	/**
-	 * Track when WordPress request starts processing
+	 * Track when WordPress init starts
 	 */
 	public static function track_request_start()
 	{
 		// Only track our API endpoint
 		if (isset($_SERVER['REQUEST_URI']) && strpos($_SERVER['REQUEST_URI'], '/wp-json/vibemap/v1/places-data') !== false) {
 			self::$request_start_time = microtime(true);
-			error_log('HRW WordPress Timing: [LIFECYCLE] WordPress init started at ' . date('H:i:s.') . substr(microtime(), 2, 3));
+
+			// Only log in debug mode
+			if (defined('WP_DEBUG') && WP_DEBUG) {
+				error_log('HRW WordPress Timing: WordPress init started for API request');
+			}
 		}
 	}
 
 	/**
-	 * Track when WordPress has loaded
+	 * Track when WordPress is fully loaded
 	 */
 	public static function track_wp_loaded()
 	{
 		if (self::$request_start_time && isset($_SERVER['REQUEST_URI']) && strpos($_SERVER['REQUEST_URI'], '/wp-json/vibemap/v1/places-data') !== false) {
 			$wp_loaded_time = round((microtime(true) - self::$request_start_time) * 1000, 2);
-			error_log('HRW WordPress Timing: [LIFECYCLE] WordPress loaded in ' . $wp_loaded_time . 'ms');
+
+			// Only log in debug mode
+			if (defined('WP_DEBUG') && WP_DEBUG) {
+				error_log('HRW WordPress Timing: WordPress loaded in ' . $wp_loaded_time . 'ms');
+			}
 		}
 	}
 
 	/**
-	 * Track REST API response serving
+	 * Track when REST response is ready to serve
+	 * 
+	 * @param bool $served Whether the request has already been served
+	 * @param WP_REST_Response $result Result to send to the client
+	 * @param WP_REST_Request $request Request used to generate the response
+	 * @param WP_REST_Server $server Server instance
+	 * @return bool
 	 */
 	public static function track_rest_serve($served, $result, $request, $server)
 	{
@@ -85,14 +100,16 @@ class HRW_WordPress_Timing
 
 		if (self::$request_start_time) {
 			$serve_time = round((microtime(true) - self::$request_start_time) * 1000, 2);
-			error_log('HRW WordPress Timing: [LIFECYCLE] REST response ready to serve in ' . $serve_time . 'ms');
 
-			// Get response size for debugging
-			if (is_a($result, 'WP_REST_Response')) {
+			// Always log REST serve time as it's critical performance data
+			error_log('HRW WordPress Timing: REST response ready in ' . $serve_time . 'ms');
+
+			// Log response size in debug mode only
+			if (defined('WP_DEBUG') && WP_DEBUG && is_a($result, 'WP_REST_Response')) {
 				$response_data = $result->get_data();
 				$json_size = strlen(json_encode($response_data));
 				$size_kb = round($json_size / 1024, 1);
-				error_log('HRW WordPress Timing: [LIFECYCLE] Response size: ' . $size_kb . 'KB');
+				error_log('HRW WordPress Timing: Response size: ' . $size_kb . 'KB');
 			}
 		}
 
@@ -100,14 +117,20 @@ class HRW_WordPress_Timing
 	}
 
 	/**
-	 * Track when WordPress request completely ends
+	 * Track when WordPress request ends
 	 */
 	public static function track_request_end()
 	{
 		if (self::$request_start_time && isset($_SERVER['REQUEST_URI']) && strpos($_SERVER['REQUEST_URI'], '/wp-json/vibemap/v1/places-data') !== false) {
 			$total_time = round((microtime(true) - self::$request_start_time) * 1000, 2);
-			error_log('HRW WordPress Timing: [LIFECYCLE] WordPress request completely finished in ' . $total_time . 'ms');
-			error_log('HRW WordPress Timing: [LIFECYCLE] ===== WORDPRESS REQUEST COMPLETE =====');
+
+			// Always log total WordPress time as it's critical for hosting delay analysis
+			error_log('HRW WordPress Timing: WordPress request finished in ' . $total_time . 'ms');
+
+			// Log performance issues
+			if ($total_time > 1000) { // More than 1 second
+				error_log('HRW WordPress Timing: SLOW WordPress processing detected: ' . $total_time . 'ms');
+			}
 
 			// Reset for next request
 			self::$request_start_time = null;
@@ -116,17 +139,15 @@ class HRW_WordPress_Timing
 
 	/**
 	 * Get timing statistics
+	 * 
+	 * @return array Timing information
 	 */
-	public static function get_timing_summary()
+	public static function get_timing_info()
 	{
 		return [
-			'message' => 'WordPress lifecycle timing is active',
-			'tracks' => [
-				'init' => 'WordPress initialization',
-				'wp_loaded' => 'WordPress fully loaded',
-				'rest_serve' => 'REST response ready to serve',
-				'shutdown' => 'WordPress request completely finished'
-			]
+			'status' => self::$request_start_time ? 'Tracking active request' : 'No active request',
+			'start_time' => self::$request_start_time,
+			'current_duration' => self::$request_start_time ? round((microtime(true) - self::$request_start_time) * 1000, 2) . 'ms' : 'N/A'
 		];
 	}
 }