… .DS_Store
… HRW_API_OPTIMIZATION_PLAN_V2.md
… assets/.DS_Store